requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
yt-dlp
transformers
torch
celery
redis
python-dotenv
google-api-python-client
boto3
prometheus-client
fastapi-users
fastapi-users[sqlalchemy]
fastapi-users[redis]
prometheus-fastapi-instrumentator
passlib
alembic
bcrypt
python-jose